Migrate the engine and legacy server to native SDL3
Replace the SDL2 and SDL2_net integration throughout the OSS tree with native SDL3 and SDL3_net APIs, without a compatibility shim or parallel backend. Port application startup, event and text input handling, rendering, surfaces, palettes, fullscreen, mouse coordinates, gamepads, timers, and resource ownership to their SDL3 contracts.

Rewrite XSocket and the legacy C++ server around SDL3_net's asynchronous address, stream, and listener objects. Preserve the engine's blocking and non-blocking receive semantics, make socket ownership movable, handle partial I/O and disconnects explicitly, and retain the existing on-wire protocol representation at serialization boundaries.

Integrate the pinned SDL3-native Clunk revision and route focus pause/resume through the sound subsystem instead of manipulating a hard-coded SDL audio device. Convert monotonic engine deadlines to 64-bit SDL3 ticks while keeping fixed-width saved and network values explicit.

Move all build targets to imported SDL3 package targets, remove obsolete SDL2 find modules and formulae, update Linux/macOS/Windows CI and the legacy-server container, preserve runtime dependency bundling, and provide the standard SDL3 Android main/static-target path. Update installation documentation accordingly.

Add regression tests for SDL3_net loopback connection, ordering, partial reads, timeouts, disconnect and move ownership; SDL event/text-buffer ownership; and compatibility of existing options, controls, gamepad masks, and representative save data. Verified with clean GCC and Clang builds, Debug/Release and ASan/UBSan builds, all CTest cases, dummy SDL startup through the menu, Docker server build/runtime dependency inspection, clang-format, YAML parsing, and an audit showing no remaining tracked SDL2 references or legacy SDL APIs.

+++ src/3d/3dobject.h
@@ -284,10 +284,10 @@ struct Object: BaseObject {
 	int mole_on;
 	int k_track;
 	int draw_mode;
-	unsigned int device_switch_latency;
+	Uint64 device_switch_latency;
 	int prev_controls;
 	int current_controls;
-	unsigned int last_send_time;
+	Uint64 last_send_time;
 
 	int helicopter;
 	int helicopter_time;
